+++ packages/@expo/cli/src/start/server/getStaticRenderFunctions.ts
@@ -27,11 +27,6 @@ function wrapBundle(str: string) {
   return str.replace(/^(__r\(.*\);)$/gm, 'module.exports = $1');
 }
 
-function stripProcess(str: string) {
-  // TODO: Remove from the metro prelude
-  return str.replace(/process=this\.process\|\|{},/m, '');
-}
-
 // TODO(EvanBacon): Group all the code together and version.
 const getRenderModuleId = (projectRoot: string): string => {
   const moduleId = resolveFrom.silent(projectRoot, 'expo-router/node/render.js');
@@ -156,14 +151,7 @@ export async function requireFileContentsWithMetro(
 
   const content = await res.text();
 
-  let bun = wrapBundle(content);
-
-  // This exposes the entire environment to the bundle.
-  if (props.environment === 'node') {
-    bun = stripProcess(bun);
-  }
-
-  return bun;
+  return wrapBundle(content);
 }
 export async function requireWithMetro<T>(
   projectRoot: string,